+++ src/main/java/it/poliba/is45/ecoshipping/dto/UtenteDto.java
@@ -18,7 +18,7 @@ import java.util.Date;
 
 public class UtenteDto {
 
-    private long idUtente;
+    private int idUtente;
     private String nome;
     private String cognome;
     private Date dataNas;
